Fix #1 - Unused vars files are deleted in new role
Unused vars files are deleted in the new role based on the
users input for the selected platforms. This is currently hardcoded
and needs to be adapted if the role template changes!

Additional minor changes include default name of the cloned dir
(ansible-role-xx) and case insensitive global regex in the templated
repository.